Restructured the source a bit.
