+++ lib/errors/base-error.js
