+++ lib/errors/invalid-job-error.js
