+++ lib/errors/warning.js
