+++ lib/errors/worker-error.js
