+++ lib/errors/worker-stop-error.js
