+++ lib/errors/worker-warning.js
