+++ test/errors/worker-stop-error.js
@@ -6,8 +6,8 @@ const describe = lab.describe
 const it = lab.test
 const expect = require('code').expect
 
-const WorkerError = require('../../errors/worker-error')
-const WorkerStopError = require('../../errors/worker-stop-error')
+const WorkerError = require('../../lib/errors/worker-error')
+const WorkerStopError = require('../../lib/errors/worker-stop-error')
 
 describe('errors', () => {
   describe('WorkerStopError', () => {
